add references and distributed mode tests

+++ lab/distributed/run.sh
@@ -1,11 +1,23 @@
 #!/bin/bash
 
-sudo clab dep -c
+sudo clab des -t distributed.clab.yaml -c
 
 docker image prune -f
 cd ../../
-docker build . -t schema-server:latest
+if [[ "$1" == "build" ]];
+then
+    docker image prune -f
+    docker build . -t schema-server:latest
+fi
 
-cd ../common
-./configure.sh distributed
-docker image prune -f
+cd lab/distributed/
+sudo clab dep -t distributed.clab.yaml -c
+
+# cd ../common
+# ./configure.sh distributed
+
+if [[ "$1" == "build" ]];
+then
+    docker image prune -f
+fi
+
